+++ recipes/datadog.rb
@@ -33,7 +33,8 @@ end
 datadog_monitor 'chef_status' do
   cookbook 'opsline-chef-client'
   init_config({
-    'status_file' => node['opsline-chef-client']['status_file']
+    'status_file' => node['opsline-chef-client']['status_file'],
+    'disabled_file' => node['opsline-chef-client']['disabled_file']
   })
   action :add
 end